refactor(desktop): split visual-smoke-fixture into per-domain modules behind a barrel
Pure move of the #1 main-process hotspot (2538 lines) into a thin
registry barrel at the original path (visual-smoke-fixture.ts, 689
lines) plus 6 per-domain seeder modules under visual-smoke/:

  - seed-helpers.ts     shared spine (NOW, session-id consts, scenario
                        sets, header/writeSession/writeJson)
  - scenarios-settings.ts   settings/connections/plan/daily-review
  - scenarios-modules.ts    skills-market + MCP
  - scenarios-artifacts.ts  ArtifactPane seed + spec writer
  - scenarios-chat.ts       turn/processing/streaming/permission/error
                        + task-ledger + live-turn projections
  - scenarios-sessions.ts   long-transcript / workstation-statuses /
                        turn-control lineage / long-sidebar / stale

The 4-symbol public surface (resolveVisualSmokeFixture,
getVisualSmokeState, seedVisualSmokeFixture, VisualSmokeFixture) stays
defined in the barrel, so the 3 non-test consumers' imports are
byte-identical. bot-onboarding-visual-smoke.ts boundary left untouched.

Zero behavior change: registry order stable, every scenario renders
identically (alignment auditor clean across all 11 fixtures).
